Create Pushingtomars.yml
im starting this project because i have faith in myself to build a foundation for my family and a future for my grandkids and keep the RomanCoinEmpire GO TO MARS! i have faith in crypto! which i would like to help support any projects that deal with anything thats with traveling to mars! LETS GOOO! Im a regular person trying to mine with ASICS and trying to switch to GPU's. I'd appreciate and love if you guys can join me in this adventure in building my project to a full crypto coin! open to listening to offers and projects!